Refactored composites import CMD
- Added CompositeRepository and Scaphold\Queries class
